add solution Syntax task 0134

+++ 1.JavaSyntax/src/com/javarush/task/task01/task0134/Solution.java
@@ -10,6 +10,6 @@ public class Solution {
     }
 
     public static long getVolume(int a, int b, int c) {
-        return a * b * c *1000;
+        return (long) a * b * c * 1000;
     }
 }
